refactor(updates): extract shared AppBanner from update/web prompts
Pull the slim top-of-app banner chrome shared by UpdateBanner and
WebUpdatePrompt into a reusable AppBanner widget (icon, message, tap, dismiss,
optional trailing actions).
